refactor payee splitting heuristics
