+++ tests/payee_desc_cases.py
@@ -1,0 +1,62 @@
+CASES = [
+    (
+        "CITY OF SAN RAMON MAYORS CONFERENCE/ DINNER FOR COUNCILMEMBERS",
+        "CITY OF SAN RAMON",
+        "MAYORS CONFERENCE/ DINNER FOR COUNCILMEMBERS",
+    ),
+    (
+        "LANER ELECTRIC SUPPLY LIGHTBULBS - LIBRARY + CITY HALL",
+        "LANER ELECTRIC SUPPLY",
+        "LIGHTBULBS - LIBRARY + CITY HALL",
+    ),
+    (
+        "MERCURIO BROS. PRINTING, INC, FD FORM",
+        "MERCURIO BROS. PRINTING, INC",
+        "FD FORM",
+    ),
+    (
+        "SHAHKAR, MEHRAN STRENGTH & BALANCE APR 2025",
+        "SHAHKAR, MEHRAN",
+        "STRENGTH & BALANCE APR 2025",
+    ),
+    (
+        "ST. FRANCIS ELECTRIC SECURING ELECTRICAL AT CREEKSIDE PATHWAY",
+        "ST. FRANCIS ELECTRIC",
+        "SECURING ELECTRICAL AT CREEKSIDE PATHWAY",
+    ),
+    (
+        "CARRION, DAMIEN TUITION REIMBURSEMENT PER MOU",
+        "CARRION, DAMIEN",
+        "TUITION REIMBURSEMENT PER MOU",
+    ),
+    (
+        "KIM, JUNG HWA KEY REFUND",
+        "KIM, JUNG HWA",
+        "KEY REFUND",
+    ),
+    (
+        "BECKER , HENRY ROPE RESCUE TECHNICIAN CLASS MAY 19-23, 2025",
+        "BECKER, HENRY",
+        "ROPE RESCUE TECHNICIAN CLASS MAY 19-23, 2025",
+    ),
+    (
+        "ARMOUR PETROLEUM SERVICE & EQUIPMENT CORP UST VISUAL INSPECTION 4/17/25",
+        "ARMOUR PETROLEUM SERVICE & EQUIPMENT CORP",
+        "UST VISUAL INSPECTION 4/17/25",
+    ),
+    (
+        "ASBURY ENVIRONMENTAL SERVICES USED MOTOR OIL RECYCLING 5/15/25",
+        "ASBURY ENVIRONMENTAL SERVICES",
+        "USED MOTOR OIL RECYCLING 5/15/25",
+    ),
+    (
+        "CINTAS CORPORATION #054 UNIFORMS UNIFORM ITEMS - PANTS FOR LAMBERTO",
+        "CINTAS CORPORATION #054 UNIFORMS",
+        "UNIFORM ITEMS - PANTS FOR LAMBERTO",
+    ),
+    (
+        "Cogent Investigations BACKGROUND",
+        "Cogent Investigations",
+        "BACKGROUND",
+    ),
+]
